% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{pu_create}
\alias{pu_create}
\title{Cria repositorio PU}
\usage{
pu_create(name, dir = ".", github_pat = NULL)
}
\arguments{
\item{name}{nome do reposit칩rio criado.}

\item{dir}{diret칩rio em que o repo ser치 criado.}

\item{github_pat}{github personal access token}
}
\description{
Cria reposit칩rio, clona, cria branch e roda \code{\link{pu_setup}}.
}

