% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{pu_setup}
\alias{pu_setup}
\title{setup power up}
\usage{
pu_setup(name, dir = ".")
}
\arguments{
\item{name}{nome do PU, que será transformado em pu."name".}

\item{dir}{pasta para criar o power-up. A última parte é o nome do PU. Recomenda-se fazer na forma pu.XXX, em camelCode.}
}
\description{
Cria o esqueleto para um power-up em uma pasta.
}

