% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{ugi}
\alias{ugi}
\title{copia de use_git_ignore}
\usage{
ugi(ignores, directory = ".", pkg = ".")
}
\arguments{
\item{ignores}{lista dos ignores a serem adicionados.}

\item{directory}{diretorios.}

\item{pkg}{objeto do tipo pacote.}
}
\description{
copia de use_git_ignore
}

