% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{pu_setup}
\alias{pu_setup}
\title{setup power up}
\usage{
pu_setup(path = ".", rstudio = TRUE)
}
\arguments{
\item{path}{pasta para criar o power-up. A última parte é o nome do power-up. Recomenda-se fazer na forma puXXX, em camelCode.}

\item{rstudio}{criar arquivo .Rproj?}
}
\description{
Cria o esqueleto para um power-up em uma pasta
}

